+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -1,4 +1,4 @@
-package racingcar.model
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
